+++ packages/gui/src/components/Editor/Controls.tsx
@@ -21,18 +21,14 @@ import { kebabCase } from 'lodash-es'
 import { useThemeProperty } from '../providers/ThemeContext'
 import { PositionInput } from '../inputs/PositionInput'
 import { TimeInput } from '../inputs/TimeInput'
+import { UnitSteps } from '../../lib'
+import { UnitRanges } from '../../data/ranges'
 
-interface ControlProps {
+interface ControlProps extends InputProps {
   field: KeyArg
 }
-const Control = ({ field }: ControlProps) => {
-  const {
-    getField,
-    getFields,
-    setField,
-    setFields,
-    removeField
-  } = useEditor()
+const Control = ({ field, ...props }: ControlProps) => {
+  const { getField, setField } = useEditor()
   const fieldset = useFieldset()
   const property = field.toString()
   const Component: ComponentType<any> = getInputComponent(property)
@@ -53,7 +49,8 @@ const Control = ({ field }: ControlProps) => {
     label: sentenceCase(property),
     themeValues: themeValues,
     keywords,
-    ...properties[property]
+    ...properties[property],
+    ...props
   }
 
   if (dependantProperties.length) {
@@ -103,9 +100,16 @@ const ComponentWithPropertyGroup = ({ dependantProperties, property, ...props }:
   )
 }
 
+type InputProps = {
+  label?: string
+  steps?: UnitSteps
+  range?: UnitRanges
+}
 export const Inputs: Record<string, any> = {}
 Object.keys(properties).forEach((field: string) => {
-  Inputs[pascal(field)] = () => <Control field={field} />
+  Inputs[pascal(field)] = (props: InputProps) => (
+    <Control {...props} field={field} />
+  )
 })
 
 type ControlsProps = {